Initial battery project start
Have setup a battery class, a random charge discharge cycle and a
function to charge control the battery. The battery will refuse to over
or undercharge. At the moment the charge controller is failing to do the
same. This is because the availableCharge and availableDischarge
attributes are not calculating correctly, this is likely as they have
been defined at the beginning, and are not passing updates of their
values as I perhaps hoped. So therefore need to make them a calculated/
updateable variable